requirements.txt
# Environment management
python-dotenv==1.0.0

# LangChain and related packages
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.13
langchain-openai==0.0.2
langgraph==0.0.17
chromadb==0.4.22

# Embeddings and Tokenizers
tiktoken==0.5.2

# Type hints
typing-extensions==4.9.0

# Search tools
tavily-python==0.3.1

# Optional: Development tools
# pytest==7.4.4
# black==23.12.1
# isort==5.13.2
